clean up some structures
