+++ crates/core/src/compiler/mod.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use datalake_compute::DatalakeComputeCompilationResults;
-use hdp_primitives::datalake::compute::DatalakeCompute;
+use hdp_primitives::datalake::DatalakeCompute;
 
 use crate::pre_processor::PreProcessorConfig;
